feat(student): Allow 'id' as URL param for TakeExam page
This change updates the `TakeExam` component to be more flexible by accepting either `submissionId` or `id` from the URL parameters.

Previously, the component was hardcoded to only use `submissionId`. This change allows the component to be used in different routing scenarios, for instance, when linking directly to an exam by its `id` before a submission has been created.

The unused `useRef` import was also removed as part of this cleanup.